\encoding{UTF-8}
\name{diffplot}
\alias{diffplot}
\title{Plot Compositional Differences}
\description{
  Make a plot showing differences and \emph{p}-values.
}

\usage{
  diffplot(comptab, col="black", plot.rect=FALSE, plot.text=TRUE)
}

\arguments{
  \item{comptab}{list or data frame, summary of comparisons generated by \code{\link{ZC_nH2O}}}
  \item{col}{the color(s) for points}
  \item{plot.rect}{plot a reference rectangle?}
  \item{plot.text}{add text labels to the points?}
}

\details{

A plot is created with points showing the differences in means of two compositional metrics: average oxidation state of carbon (\ZC) and water demand per residue in formation reactions from basis species (\nH2O{&#772}).

For each dataset, the point symbol is a filled square if the \emph{p}-values of both \ZC and \nH2O{&#772;} are less than 0.05, a filled circle if the \emph{p}-value of one of \ZC or \nH2O{&#772;} is less than 0.05, and an open circle otherwise.

A solid line is drawn from the point to the corresponding axis if the rounded, absolute value of (\code{\link{CLES}} in percent - 50) of \ZC or \nH2O{&#772} is greater than or equal 10.
Otherwise, a dashed line is drawn from the point to the corresponding axis if the \emph{p}-value of \ZC or \nH2O{&#772;} is less than 0.05.
Otherwise, no line is drawn.

The colors of the points are controlled by \code{col}, which is recycled to be equal to the number of comparisons in \code{comptab}.

If \code{plot.rect} is TRUE, a shaded \code{\link{rect}}angle is drawn with coordinates -0.01, -0.01, 0.01, 0.01.
This is useful for emphasizing the different scales of adjacent plots.

If \code{plot.text} is TRUE, the points are plotted in a bigger size and \code{\link{text}} labels are added.
The labels are lowercase letters taken from the Latin, then Greek alphabets.

}

\examples{
data(canprot)
comptab <- lapply(c("JKMF10", "WDO+15_C.N"), function(dataset) {
  pdat <- get_pdat(dataset)
  ZC_nH2O(pdat, plot.it=FALSE)
})
diffplot(comptab)
}

\concept{Chemical composition}
