\encoding{UTF-8}
\name{diffplot}
\alias{diffplot}
\title{Plot Compositional Differences}
\description{
  Make a plot showing differences and \emph{p}-values.
}

\usage{
  diffplot(comptab, col="black")
}

\arguments{
  \item{comptab}{list or data frame, summary of comparisons generated by \code{\link{ZC_nH2O}}}
  \item{col}{the color(s) for points}
}

\details{

A plot is created with points showing the differences in means of two compositional metrics: average oxidation state of carbon (\ZC) and water demand per residue in formation reactions from basis species (\nH2O{&#772}).

For plotting the points, if either the absolute value of (\code{\link{CLES}} in percent - 50) is greater than 10 or \emph{p}-value is less than 0.05, it is considered a \dQuote{significant} difference.
The point symbol is an open circle if neither (\ZC) nor (\nH2O{&#772;}) has a significant difference, a filled circle if one of the metrics has a significant difference, and a filled square if both metrics have a significant difference.
The colors of the points are controlled by \code{col} (recycled to be equal to the number of comparisons in \code{comptab}).

For each metric that has a significant difference, a line is drawn from its point that is normal to and ends at the corresponding axis line.
The line is solid if both the effect size and \emph{p}-value meet the criteria above, or dashed line if only one of effect size or p-value meets the criterion.
}

\examples{
data(canprot)
comptab <- lapply(c("JKMF10", "WDO+15_C.N"), function(dataset) {
  pdat <- get_pdat(dataset)
  ZC_nH2O(pdat, plot.it=FALSE)
})
diffplot(comptab)
}

\concept{Chemical composition}
