\encoding{UTF-8}
\name{pdat_}
\alias{pdat_}
\alias{pdat_CRC}
\alias{pdat_pancreatic}
\alias{pdat_hypoxia}
\alias{pdat_osmotic}
\alias{pdat_osmotic2}
\alias{pdat_multi}
\alias{pdat_secreted}
\alias{pdat_breast}
\title{Get Protein Expression Data}
\description{
  Get data on protein expression and chemical composition.
}

\usage{
  pdat_CRC(dataset = 2020, basis = "rQEC")
  pdat_pancreatic(dataset = 2020, basis = "rQEC")
  pdat_breast(dataset = 2020, basis = "rQEC")
  pdat_hypoxia(dataset = NULL, basis = "rQEC")
  pdat_osmotic(dataset = NULL, basis = "rQEC")
  pdat_osmotic2(dataset = NULL, basis = "rQEC")
  pdat_multi(dataset = 2020, basis = "rQEC")
  pdat_secreted(dataset = 2020, basis = "rQEC")
}

\arguments{
  \item{dataset}{character, dataset name}
  \item{basis}{character, keyword for basis species to use}
}

\details{
The \code{pdat_} functions calculate chemical compositional metrics (using \code{\link{protcomp}}) for relatively up- and down-expressed proteins reported in proteomic experiments.

Use \code{pdat_CRC}, \code{pdat_pancreatic}, and \code{pdat_breast} to retrieve data for protein expression in different cancer types, \code{pdat_hypoxia} for data on hypoxia or 3D culture, and \code{pdat_osmotic} for data on hyperosmotic stress.
\code{pdat_secreted} retrieves data for secreted proteins, including exosomes, in hypoxia.
\code{pdat_multi} has data from studies that include multiple conditions (e.g. cellular and secreted proteins in hypoxia), and is called by the corresponding functions (\code{pdat_hypoxia}, \code{pdat_secreted}).
The functions get relative expression data from the CSV files stored in \code{extdata/expression/}, with subdirectories corresponding to the names of the functions.
Some of the functions also retrieve amino acid compositions from the files in \code{extdata/aa/} (for non-human proteins).

\code{pdat_osmotic2} adds more data for osmotic stress experiments on bacteria and archaea, as described in Dick et al. (2019).

If \code{dataset} is \code{NULL}, the function returns the names of all datasets in the compilation.
If \code{dataset} is \samp{2017} or \samp{2017}, it indicates the year of the compilation.

Provide one of the dataset names as the \code{dataset} argument to retrieve the data.
Each dataset name indicates the study (publication) where the data were reported, constructed by combining the first characters of the (first three or four) authors' family names with the 2-digit year of publication.
This coincides with the key-generation scheme used in some bibliography manager software.
This abbreviation is also used to name the CSV file containing the data.
If more than one dataset is available from a single study (for example, for relative protein expression in different stages of cancer), \code{dataset} an underscore followed by a short suffix indicating the particular dataset.

Tables listing mean compositional differences between up- and down-expressed proteins for each dataset are saved in \code{extdata/summary/}.
These files were created using the second example below.
}

\value{
A list consisting of \code{dataset} (the name of the dataset), \code{basis} (basis species used for the calculations), \code{description} (descriptive text), \code{pcomp} (compositional data generated by \code{\link{protcomp}}), \code{up2} (logical vector with length equal to the number of proteins; TRUE if the protein is up-expressed in group 2 compared to group 1 (i.e. cancer compared to normal), FALSE otherwise).
}

\references{
Dick, J. M., Tan, J. and Yu, M. (2019) Geochemistry of metagenomes: Redox and salinity gradients and the chemical compositions of proteins. \emph{(preprint)}
}

\examples{
library(CHNOSZ)
# list datasets in 2017 complilation
pdat_CRC(2017)
pdat_CRC("JKMF10")

\dontrun{
# how the extdata/summary/summary_*.csv files were made
for(what in c("CRC", "pancreatic", "hypoxia", "osmotic")) {
  pdat_fun <- paste0("pdat_", what)
  datasets <- get(pdat_fun)()
  comptab <- lapply_canprot(datasets, function(dataset) {
    pdat <- get(pdat_fun)(dataset)
    get_comptab(pdat)
  }, varlist = "pdat_fun")
  # write summary table
  comptab <- do.call(rbind, comptab)
  comptab <- cbind(set = c(letters, LETTERS)[1:nrow(comptab)], comptab)
  comptab[, 6:15] <- signif(comptab[, 6:15], 4)
  filename <- paste0("summary_", what, ".csv")
  write.csv(comptab, filename, row.names = FALSE, quote = 3)
}}
}

\concept{Protein expression}
