\encoding{UTF-8}
\name{pdat_}
\alias{pdat_}
\alias{pdat_CRC}
\alias{pdat_pancreatic}
\alias{pdat_hypoxia}
\alias{pdat_osmotic}
\title{Get Protein Data}
\description{
  Get data on protein expression and chemical composition.
}

\usage{
  pdat_CRC(dataset = NULL, basis = "AA")
}

\arguments{
  \item{dataset}{character, specifies which dataset to retrieve}
  \item{basis}{character, keyword for basis species to use}
}

\details{
The \code{pdat_} functions read CSV data files for data on relatively up- and down-expressed proteins reported in proteomic experiments, then uses \code{\link{protcomp}} to calculate some chemical compositional metrics.

\code{pdat_CRC} retrieves data for protein expression in colorectal cancer.
\code{pdat_pancreactic} retrieves data for protein expression in pancreatic cancer.
\code{pdat_hypoxia} retrieves data for protein expression in hypoxia or 3D culture.
\code{pdat_osmotic} retrieves data for protein expression in hyperosmotic stress.

If \code{dataset} is \code{NULL}, the return value gives the names of all datasets that can be retrieved using this function.
Each name indicates the study (publication) where the data were reported, constructed by combining the first characters of the (first three or four) authors' family names with the 2-digit year of publication.
This coincides with the key-generation scheme used in some bibliography manager software.
This abbreviation also corresponds the name of the CSV file containing the data.
If more than one dataset is available from a single study (for example, for relative protein expression in different stages of cancer), \code{dataset} is suffixed by an underscore followed by a short abbreviation indicating the particular dataset.

Provide one of these names as the \code{dataset} argument to retrieve the data.
}

\value{
A list consisting of \code{dataset} (the name of the dataset), \code{basis} (basis species used for the calculations), \code{description} (descriptive text), \code{pcomp} (compositional data generated by \code{\link{protcomp}}), \code{up2} (logical vector with length equal to the number of proteins; TRUE if the protein is up-expressed in group 2 compared to group 1 (i.e. cancer compared to normal), FALSE otherwise), \code{names} (gene names for the proteins, if available).
}

\seealso{
  \code{\link{get_pdat}}
}

\examples{
data(canprot)
pdat_CRC()
pdat_CRC("JKMF10")  # same result as get_pdat("JKMF10")
}

\concept{Protein expression}
