\encoding{UTF-8}
\name{pdat_}
\alias{pdat_}
\alias{pdat_breast}
\alias{pdat_colorectal}
\alias{pdat_lung}
\alias{pdat_liver}
\alias{pdat_pancreatic}
\alias{pdat_prostate}
\alias{pdat_hypoxia}
\alias{pdat_secreted}
\alias{pdat_3D}
\alias{pdat_glucose}
\alias{pdat_osmotic_bact}
\alias{pdat_osmotic_euk}
\alias{pdat_halophiles}
\alias{.pdat_multi}
\alias{.pdat_osmotic}
\title{Get Protein Expression Data}
\description{
  Get data on protein expression and chemical composition.
}

\usage{
  pdat_breast(dataset = 2020, basis = "rQEC")
  pdat_colorectal(dataset = 2020, basis = "rQEC")
  pdat_liver(dataset = 2020, basis = "rQEC")
  pdat_lung(dataset = 2020, basis = "rQEC")
  pdat_pancreatic(dataset = 2020, basis = "rQEC")
  pdat_prostate(dataset = 2020, basis = "rQEC")
  pdat_hypoxia(dataset = 2020, basis = "rQEC")
  pdat_secreted(dataset = 2020, basis = "rQEC")
  pdat_3D(dataset = 2020, basis = "rQEC")
  pdat_glucose(dataset = 2020, basis = "rQEC")
  pdat_osmotic_bact(dataset = 2020, basis = "rQEC")
  pdat_osmotic_euk(dataset = 2020, basis = "rQEC")
  pdat_halophiles(dataset = 2020, basis = "rQEC")
  .pdat_multi(dataset = 2020, basis = "rQEC")
  .pdat_osmotic(dataset = 2017, basis = "rQEC")
}

\arguments{
  \item{dataset}{character, dataset name}
  \item{basis}{character, keyword for basis species to use}
}

\details{
The \code{pdat_} functions calculate chemical compositional metrics (using \code{\link{protcomp}}) for relatively up- and down-expressed proteins reported in proteomic experiments.

Use \code{pdat_colorectal}, \code{pdat_pancreatic}, \code{pdat_breast}, \code{pdat_lung}, \code{pdat_prostate}, and \code{pdat_liver} to retrieve data for protein expression in different cancer types
Use \code{pdat_hypoxia} for data on hypoxia in cell culture and \code{pdat_glucose} for high-glucose experiments.
\code{pdat_3D} retrieves data for 3D vs 2D cell culture (including tumor spheroids) and \code{pdat_secreted} retrieves data for secreted proteins (including exosomes) in hypoxia.

\code{.pdat_osmotic} retrieves data for hyperosmotic stress, for the 2017 compilation only.
In 2020, these datasets were split among \code{pdat_osmotic_bact} (bacteria) \code{pdat_osmotic_euk} (eukaryotes) and \code{pdat_halophiles} (halophilic bacteria and archaea), as described in Dick et al. (2020).

\code{.pdat_multi} retrieves data from studies that include multiple conditions (e.g. cellular and secreted proteins in hypoxia), and is used internally by the specific functions (\code{pdat_hypoxia}, \code{pdat_secreted}).

The functions get relative expression data from the CSV files stored in \code{extdata/expression/}, with subdirectories corresponding to the names of the functions.
Some of the functions also retrieve amino acid compositions from the files in \code{extdata/aa/} (for non-human proteins).

If \code{dataset} is \code{NULL}, the function returns the names of all datasets in the compilation.
If \code{dataset} is \samp{2020} (the default) or \samp{2017}, it indicates the year of the compilation.

Provide one of the dataset names as the \code{dataset} argument to retrieve the data.
Each dataset name indicates the study (publication) where the data were reported, constructed by combining the first characters of the (first three or four) authors' family names with the 2-digit year of publication.
This coincides with the key-generation scheme used in some bibliography manager software.
This abbreviation is also used to name the CSV file containing the data.
If more than one dataset is available from a single study (for example, for relative protein expression in different stages of cancer), \code{dataset} an underscore followed by a short suffix indicating the particular dataset.
}

\value{
A list consisting of \code{dataset} (the name of the dataset), \code{basis} (basis species used for the calculations), \code{description} (descriptive text), \code{pcomp} (compositional data generated by \code{\link{protcomp}}), \code{up2} (logical vector with length equal to the number of proteins; TRUE if the protein is up-expressed in group 2 compared to group 1 (i.e. cancer compared to normal), FALSE otherwise).
}

\references{
Dick, J. M., Yu, M. and Tan, J. (2020) Distinct trends in chemical composition of proteins from metagenomes in redox and salinity gradients. \emph{bioRxiv}. \doi{10.1101/2020.04.01.020008}
}

\examples{
library(CHNOSZ)
# list datasets in 2017 complilation for colorectal cancer
pdat_colorectal(2017)
# process one dataset
pdat_colorectal("JKMF10")
}

\concept{Protein expression}
